fix: implement correct actualization

+++ 04-copy-directory/index.js
@@ -10,13 +10,15 @@ const copyDir = function () {
   });
 
   fs.readdir(path.join(__dirname, 'files-copy'), function (err, files) {
-    files.forEach(item => {
-      fs.unlink(path.join(__dirname, 'files-copy', item), function (err) {
-        if (err) {
-          return console.log(err);
-        }
-      })
-    });
+    if (Array.isArray(files)) {
+      files.forEach(item => {
+        fs.unlink(path.join(__dirname, 'files-copy', item), function (err) {
+          if (err) {
+            return console.log(err);
+          }
+        })
+      });
+    }
   });
 
   fs.readdir(path.join(__dirname, 'files'), function (err, files) {
